backend_ai/requirements.txt
fastapi
uvicorn
firebase-admin
google-cloud-firestore
google-generativeai
python-dotenv
